build: rebuild dist and types after the #47 de-duplication
One-shot artifact rebuild for the integration of the stub/monolith
de-duplication (#62-#70) into dev, per the agreed convention (refactor PRs
ship source only; artifacts are rebuilt at integration/release points).

- dist/jsfeatNext.js (UMD) + dist/jsfeatNext.mjs (ESM) rebuilt from the
  refactored source
- types/ regenerated: now mirrors the new module layout, including
  types/src/core/core.d.ts and types/src/motion_model/motion_model.d.ts

Verified: npm test 57/57; all 18 public modules attach on the fresh UMD
bundle; VERSION 0.7.6 intact; no .d.ts files leak into dist/.
